[FIX] Correcciones en métodos para asignación de etiquetas analíticas
