+++ fertico-addons/models/__init__.py
@@ -3,7 +3,7 @@
 from . import models
 from . import account_analytic
 from . import account_invoice_correction
-from . import analytic_tag_new_control
+from . import analytic_tag_analysis
 from . import attendances
 from . import contact
 from . import domain_terms
